Auth

+++ core/app/controllers/Application.java
@@ -54,12 +54,15 @@ public class Application extends Controller {
         return configList;
     }
 
-    public Result index() {
+    // eventually things should all be seperated into a class, and this should be at the class level
+    @Security.Authenticated(Secured.class)
+    public Result index() {        
         IndexViewModel viewModel = new IndexViewModel();
         viewModel.configurations = getConfigurations();
         return ok(index.render(viewModel));
     }
 
+    @Security.Authenticated(Secured.class)
     public Result addConfiguration() {
         AddConfigurationViewModel viewModel = new AddConfigurationViewModel();
 
@@ -84,13 +87,19 @@ public class Application extends Controller {
         return ok(addConfiguration.render(viewModel));
     }
 
+    @Security.Authenticated(Secured.class)
     public Result deleteConfiguration() {
         DynamicForm bindedForm = new DynamicForm().bindFromRequest();
         FrontEndDBInterface f = new FrontEndDBInterface();
-        f.deleteConfigAndRecords(Integer.parseInt(bindedForm.get("conf")));
+        String conf_id = bindedForm.get("conf");
+        if (!Secured.canAccess(request().username(), conf_id)) {
+            return this.authError();
+        }
+        f.deleteConfigAndRecords(Integer.parseInt(conf_id));
         return redirect("/");
     }
 
+    @Security.Authenticated(Secured.class)
     public Result submitConfiguration() {
         DynamicForm bindedForm = new DynamicForm().bindFromRequest();
         FrontEndDBInterface f = new FrontEndDBInterface();
@@ -109,7 +118,11 @@ public class Application extends Controller {
         return redirect("/");
     }
 
+    @Security.Authenticated(Secured.class)
     public Result configuration(String configuration_id) {
+        if (!Secured.canAccess(request().username(), configuration_id)) {
+            return this.authError();
+        }
         RecipeViewModel viewModel = new RecipeViewModel();
         FrontEndDBInterface f = new FrontEndDBInterface();
         models.Configuration conf;
@@ -127,13 +140,17 @@ public class Application extends Controller {
         return ok(recipe.render(viewModel));
     }
 
+    @Security.Authenticated(Secured.class)
     public Result addRun(String configuration_id) {
+        if (!Secured.canAccess(request().username(), configuration_id)) {
+            return this.authError();
+        }
         AddRunViewModel viewModel = new AddRunViewModel();
 
         // should also get name passed through here
         viewModel.configuration_id = configuration_id;
         viewModel.default_url = "";
-        viewModel.default_author = "";
+        viewModel.default_author = request().username();
         viewModel.default_repo = "";
         viewModel.default_comment = "";
         viewModel.error_message = "";
@@ -141,10 +158,15 @@ public class Application extends Controller {
         return ok(addRun.render(viewModel));
     }
 
+    @Security.Authenticated(Secured.class)
     public Result submitRun() {
         DynamicForm bindedForm = new DynamicForm().bindFromRequest();
 
         String conf_id = bindedForm.get("configuration_id");
+
+        if (!Secured.canAccess(request().username(), conf_id)) {
+            return this.authError();
+        }
 
         String url = bindedForm.get("url");
         String author = bindedForm.get("author");
@@ -168,11 +190,9 @@ public class Application extends Controller {
         String record_id = f.storeRunInfo(Integer.parseInt(conf_id), url, author, repo, comment);
 
         masterActor.tell(new SetUpJobMessage(conf_id, url, record_id), masterActor);
-        WorkingViewModel viewModel = new WorkingViewModel();
-        viewModel.record_id = record_id;
-        viewModel.conf_id = conf_id;
-        viewModel.percent_complete = 0;
-        return ok(working.render(viewModel));
+
+        return redirect("/record?record_id="+record_id);
+
     }
 
     public Result progressBar(String record_id, String conf_id) {
@@ -185,45 +205,55 @@ public class Application extends Controller {
 
     public Promise<Result> getCurrentProgress(String record_id) {
         return Promise.wrap(ask(masterActor, new StatusRequest(record_id), 60000))
-                .map(new Function<Object, Result>() {
-                    public Result apply(Object response) {
-                        ObjectNode result = Json.newObject();
-                        if (response instanceof StatusUpdate) {
-                            StatusUpdate update = ((StatusUpdate) response);
-                            int percentComplete;
-                            if (update.getTotal() > 0) {
-                                double comp = ((double) (update.getCompleted() + update.getSkipped())) / ((double) update.getTotal());
-                                percentComplete = (int) (comp * 100.0);
-                            } else {
-                                percentComplete = 0;
-                            }
-                            result.put("percent_complete", Integer.toString(percentComplete));
-                            result.put("completed", Integer.toString(update.getCompleted()));
-                            result.put("skipped", Integer.toString(update.getSkipped()));
-                            result.put("total", Integer.toString(update.getTotal()));
-                            return ok(result);
+            .map(new Function<Object, Result>() {
+                public Result apply(Object response) {
+                    ObjectNode result = Json.newObject();
+                    if (response instanceof StatusUpdate) {
+                        StatusUpdate update = ((StatusUpdate) response);
+                        int percentComplete;
+                        if (update.getTotal() > 0) {
+                            double comp = ((double) (update.getCompleted() + update.getSkipped())) / ((double) update.getTotal());
+                            percentComplete = (int) (comp * 100.0);
+                        } else {
+                            percentComplete = 0;
                         }
-                        result.put("percent_complete", "0");
-                        result.put("completed", "0");
-                        result.put("skipped", "0");
-                        result.put("total", "0");
+                        result.put("percent_complete", Integer.toString(percentComplete));
+                        result.put("completed", Integer.toString(update.getCompleted()));
+                        result.put("skipped", Integer.toString(update.getSkipped()));
+                        result.put("total", Integer.toString(update.getTotal()));
                         return ok(result);
                     }
-                });
+                    result.put("percent_complete", "0");
+                    result.put("completed", "0");
+                    result.put("skipped", "0");
+                    result.put("total", "0");
+                    return ok(result);
+                }
+                
+            });
     }
 
+    @Security.Authenticated(Secured.class)
     public Result record(String record_id) {
         RecordViewModel viewModel = new RecordViewModel();
         FrontEndDBInterface f = new FrontEndDBInterface();
         Record associated_record = f.getRecordFromRecordID(Integer.parseInt(record_id));
 
+        if (!Secured.canAccess(request().username(), associated_record.configuration_id)) {
+            return this.authError();
+        }
+
         viewModel.record = associated_record;
         return ok(record.render(viewModel));
     }
 
+    @Security.Authenticated(Secured.class)
     public Result deleteRecord() {
         DynamicForm bindedForm = new DynamicForm().bindFromRequest();
         String conf_id = bindedForm.get("configuration_id");
+        if (!Secured.canAccess(request().username(), conf_id)) {
+            return this.authError();
+        }
         FrontEndDBInterface f = new FrontEndDBInterface();
         f.deleteRecordFromRecordID(Integer.parseInt(bindedForm.get("record_id")));
         return redirect("/configuration?conf=" + conf_id);
@@ -231,5 +261,99 @@ public class Application extends Controller {
 
     public Result about() {
         return ok(about.render());
+    }
+
+    private List<String> getTeamNames() {
+        List<String> teamNames = new ArrayList<>();
+
+        teamNames.add("team NN1");
+        teamNames.add("team NN2");
+        teamNames.add("team CRF1");
+        teamNames.add("team CRF2");
+        teamNames.add("team Perc1");
+        teamNames.add("team Perc2");
+        teamNames.add("team SVM1");
+        teamNames.add("team SVM2");
+        teamNames.add("team Exp1");
+        teamNames.add("team Exp2");
+
+        return teamNames;
+    }
+
+    public Result login() {
+        LoginViewModel viewModel = new LoginViewModel();
+
+        viewModel.teamNames = getTeamNames();
+
+        return ok(login.render(viewModel));
+    }
+
+    public Result addUser() {
+        DynamicForm bindedForm = new DynamicForm().bindFromRequest();
+
+        String username = bindedForm.get("username");
+        String verify = bindedForm.get("verify");
+        String password = bindedForm.get("password");
+        String teamName = bindedForm.get("teamName");
+
+        LoginViewModel viewModel = new LoginViewModel();
+        viewModel.username = username;
+        viewModel.teamNames = getTeamNames();
+        viewModel.teamName = teamName;
+        if(! password.equals(verify)) {
+            String error = "Password and password verification do not match"; 
+            viewModel.errorMessage = error;
+            return ok(login.render(viewModel));
+        } else if (password.length() == 0) {
+            String error = "Password cannot be of length 0";
+            viewModel.errorMessage = error;
+            return ok(login.render(viewModel));
+        }
+
+        boolean teamPassCorrect = true;
+        //@Deepak add check that the team password is correct
+        if (!teamPassCorrect) {
+            String error = "Team password is incorrect";
+            viewModel.errorMessage = error;
+            return ok(login.render(viewModel));
+        }
+
+
+        //@Deepak insert a record into the user db with this username and pw
+        
+        session().clear();
+        session("username", username);
+        return redirect("/");
+    }
+
+    public Result loginUser() {
+        DynamicForm bindedForm = new DynamicForm().bindFromRequest();
+
+        String username = bindedForm.get("loginUsername");
+        String password = bindedForm.get("loginPassword");
+
+        //@Deepak get password for username from db
+        if(/*some error occues*/ password.equals("errorCheck")) {
+            String error = "Some error"; 
+            LoginViewModel viewModel = new LoginViewModel();
+            viewModel.errorMessage = error;
+            viewModel.loginUsername = username;
+            viewModel.teamNames = getTeamNames();
+            return ok(login.render(viewModel));
+        }
+        
+        session().clear();
+        session("username", username);
+        return redirect("/");
+    }
+
+    @Security.Authenticated(Secured.class)
+    public Result logout() {
+        session().clear();
+        return redirect("/");
+    }
+
+    private Result authError() {
+        return forbidden("403 Forbidden: User does not have access to this configuration");
     }
 }
